fix: Remove the Y/N option from the form
Redundant Y/N inputs have been removed from both
Company and Partnership business types. Instead,
the business type is set when the button is pressed
to change the business type being entered.
Side note: This change means that when the business
type is changed the form fields and state are reset.