feat: Change 'Create Variation' button to 'Save As Draft'
When a variation is created it is first saved as a
draft and must then be edited before being able to
be sent to the Owner for sign off. In order to add
a button to send the form straight to the Owner to
sign off for approval during the creation process
a 'Save As Draft' button has been added so that a
'Send Variation For Approval' button can also be
added. This differentiates creating a Variation
that can be edited later from one we want to send
directly to the owner, bypassing the 'draft' step.