feat: Redirect user to variation detail page after creation
When a variation is created the variation is
stored in the database and the id of the newly created
variation is returned to the front end. The user is now
directed to the details page for the new variation
as opposed to returning to the project page. This
will easier enable the variation to be submitted for
review in a future commit as the logic already exists
on the 'Send Variation For Approval Button'.
The decision to do it this way was because it was
very difficult to create logic for sending a variation
for approval during the creation process as the variation
doesn't exist. As a side note - the Save Draft commit
I made earlier is now defunct, and the button has been
renamed back to 'Create Variation'.